Latest stable version of "GitHub Starred Repositories" app.
This version of the app, includes a new feature, the user can clickan item so that he can check the repository he selected, this is done by putting an extra -that contains contains the URL of teh repository- to the intent which start a new activity that contains a WebView that loads the URL of the selected repository.
This version contains also some improvments in UI.
v. 1.3.1.2